requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
blinker==1.4
boto3==1.5.36
botocore==1.8.50
celery==3.1.18
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.8.2
django-allauth==0.39.1
django-bootstrap-form==3.3
django-debug-toolbar==1.11
django-decouple==2.1
django-extensions==2.1.6
django-storages==1.6.5
dnspython==1.16.0
docutils==0.14
eventlet==0.24.1
falcon==1.4.1
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
jmespath==0.9.3
kombu==3.0.37
mongoengine==0.17.0
monotonic==1.5
oauthlib==3.0.1
Pillow==5.0.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pydotplus==2.0.2
pymongo==3.7.2
pyparsing==2.4.0
python-dateutil==2.6.1
python-decouple==3.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2017.3
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.1.13
six==1.11.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.24.1
Werkzeug==0.15.2
